refactor(frontend): move WorkspaceCommandListProvider to its own file
It was living in workspace-emoji.tsx because both are workspace-metadata
providers, but mixing unrelated concepts in one file is confusing.
