+++ apps/frontend/src/components/workspace-emoji.tsx
@@ -1,8 +1,6 @@
-import { useMemo, type ReactNode } from "react"
+import type { ReactNode } from "react"
 import { useWorkspaceEmoji } from "@/hooks/use-workspace-emoji"
-import { useWorkspaceMetadata } from "@/stores/workspace-store"
 import { EmojiProvider } from "@/lib/markdown/emoji-context"
-import { CommandListProvider } from "@/lib/markdown/command-list-context"
 
 interface WorkspaceEmojiProps {
   workspaceId: string
@@ -32,19 +30,4 @@ interface WorkspaceEmojiProviderProps {
 export function WorkspaceEmojiProvider({ workspaceId, children }: WorkspaceEmojiProviderProps) {
   const { emojis } = useWorkspaceEmoji(workspaceId)
   return <EmojiProvider emojis={emojis}>{children}</EmojiProvider>
-}
-
-interface WorkspaceCommandListProviderProps {
-  workspaceId: string
-  children: ReactNode
-}
-
-/**
- * Provides the registered slash command names for rendering, so "/foo" in
- * message text is only styled as a command chip when `foo` is a real command.
- */
-export function WorkspaceCommandListProvider({ workspaceId, children }: WorkspaceCommandListProviderProps) {
-  const metadata = useWorkspaceMetadata(workspaceId)
-  const commandNames = useMemo(() => metadata?.commands?.map((c) => c.name) ?? [], [metadata?.commands])
-  return <CommandListProvider commandNames={commandNames}>{children}</CommandListProvider>
 }
